m/n/core/consensus/client: implement ThinClient()
Summary:
This implements a ThinClient function on Namespaced which returns an
clientv3.Client. These can be useful when dealing with existing code
which expects this type, but only uses non-management APIs. For example,
the clientv3 concurrency library.

Test Plan: To be used by future code, and basic enough that it IMO does not warrant unit tests?

Change-Id: Ic6e38ff654bafd8bb385cf108637f8ed058015dc
Reviewed-on: https://review.monogon.dev/c/monogon/+/121
Reviewed-by: Lorenz Brun <lorenz@nexantic.com>