m/test/launch: fix TPM tempdir permissions, wrap errors
On Linux, the following generally fails:

    $ cd /tmp
    $ mkdir test
    $ cd test/
    $ chmod 644 .
    $ touch foo
    touch: cannot touch 'foo': Permission denied

This changes our launch code to create a temporary TPM directory with
755 instead of 644 permissions, preventing a situation like above
manifesting in our new CI.

This didn't manifest before as we always ran builds through podman, and
there this behaviour doesn't appear to hold, probably because we are uid
0 there:

    $ podman exec -it monogon-dev bash
    bash-5.0# id
    uid=0(root) gid=0(root) groups=0(root) context=unconfined_u:system_r:spc_t:s0
    bash-5.0# cd /tmp/
    bash-5.0# mkdir test
    bash-5.0# cd test/
    bash-5.0# chmod 644 .
    bash-5.0# touch foo

We also drive-by some unwrapped error returns to be a bit more helpful.

Test Plan: Tested on new CI, manually.

X-Origin-Diff: phab/D773
GitOrigin-RevId: 5a55a7878109717f0c17251a659dfc6ee04b94f4