third_party/go: add package missing from dependency graph
This is a Windows-specific package being pulled in by github.com/spf13/cobra.

We don't need it, and we don't ever build it (it's behind a select()
gate depending on the Windows platform), but its lack causes us to not
be able to perform Bazel queries against anything that stumbles upon
this select statement.

Notably, things like ibazel don't work without the ability to query
dependencies of a target. In theory, cquery could be used of query (and
cquery would know that it's not running on a windows platform and not
attempt to resolve the missing package). This might happen some day,
but:

  1) cquery currently does not support the buildfiles() function, which
     is needed by tools like ibazel to find not only source/data/target
     dependencies for a taret, but also every BUILD/.bzl file that
     influenced that target.

     See: https://github.com/bazelbuild/bazel-watcher/issues/305#issuecomment-627312885

  2) It's generally good practice to not have missing objects in our
     dependency graph, I think. We will sooner or later start using this
     data in CI and other automation, and it might be useful to make an
     assumption, at some point, that we don't ever have a broken
     target dependency graph.

Testing plan: the following now works:

   bazel query 'deps(set(//...))' --output=xml

Change-Id: Ic45e293b868b0aaa707f31384b4b24626ba23e29
Reviewed-on: https://review.monogon.dev/c/monogon/+/200
Reviewed-by: Leopold Schabel <leo@nexantic.com>