handbook: structure, introduction, chapter 3 root
This is a first pass at the Metropolis Handbook. We write out the
expected outline of the documentation, and start filling in some
chapters.

The introduction pages and Chapter 1 are a high-level overview of what
Metropolis is and how it's expected to be used.

Chapter 3 will delve deep into the technical details of how Metropolis
is implemented. For now, the introduction/overview section of that
chapter is written.

Change-Id: Ic16aa91ed9a127f4f3791eeaf8d7f5e1a24b1018
Reviewed-on: https://review.monogon.dev/c/monogon/+/453
Tested-by: Jenkins CI
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>