m/p/devicemapper: fix GC closing control fd
The devicemapper package stored a reference to its control file
descriptor as a uintptr after opening it thorugh os.Open(). This is a
problem as os.newFile (internally called by os.Open) sets a finalizer
on the os.File which closes the fd as soon as the object is GCed.
Because no such reference was kept by the devicemapper package, the GC
could end up closing the fd.

To fix this, the package now keeps the original os.File around and
just grabs an Fd as necessary.  While we're at it, let's make the
control file descriptor implementation threadsafe.

Change-Id: I6f7e0a398f28c1141627904ccbd2d99dd248bc78
Reviewed-on: https://review.monogon.dev/c/monogon/+/310
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>
Vouch-Run-CI: Lorenz Brun <lorenz@monogon.tech>