m/n/c/curator: return CA public key in GetClusterInfo
This is needed for node registration (and is generally useful data
whenever a caller might not be aware of the CA's public key but already
has access to a Management client). In theory, all callers should be
aware of the public key, but in the future some other cluster
verification might be performed with the CA public key ignored on
connectivity, but used by some other logic.

Change-Id: If1928435bd5606c733460eb1a4a29a6578c8c723
Reviewed-on: https://review.monogon.dev/c/monogon/+/342
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>