m/pkg/event/etcd: implement ranged watchers
This adds a new mode of operation to etcd Values/Watchers in which a
range of etcd keys is watched for updates instead of a single key.

This allows the implementation of watching a collection of objects
stored in etcd for updates, eg. the node state in the Curator.

This has been implemented within the existing API of Event Values, which
is likely the biggest contention point of this change. An alternative
would be to design a separate API for multi-value use, but this should
allow us to more easily integrate with the existing code. We make use of
Go's options-as-varargs paradigm to not break any existing use of this
codebase.

Some behaviour of the Get() operation in ranged context is left
underdefined, but none of the expected users of this codebase are
expected to depend on this. Once the dust settles a bit, we can attempt
to formalize this more strongly.

Change-Id: I8f84d74332765e52b9bbec04b626d00f05c23071
Reviewed-on: https://review.monogon.dev/c/monogon/+/419
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>