m/c/metroctl/core: add frontend-independent metroctl support pkg
This adds metroctl/core, a package which contains parts of metroctl
which do significant amounts of work beyond just providing a CLI for
them.
This package is intended to be used for integrating with functions
provided by metroctl, for example for using them in integration tests
or writing other frontends providing functionality similar to metroctl
(like a GUI or webapp).

Change-Id: I8a56bfbefce8d18c6c9be3349e3c7a15a699d009
Reviewed-on: https://review.monogon.dev/c/monogon/+/411
Reviewed-by: Mateusz Zalega <mateusz@monogon.tech>
Vouch-Run-CI: Mateusz Zalega <mateusz@monogon.tech>