m/n/c/network/hostsfile: implement
This implementes a dedicated runnable for maintaining hostsfile-like
local state based on the node's local state and any possible cluster
data.

This needs to be able to be maintained by a single runnable regardless
of the cluster enrolment process (bootstrap, register or join), and
regardless of the state of enrolment (don't have networking data, only
have local networking data, have cluster state from any kind of
available cluster dialer).

For now this is just piped into the bootstrap logic and has no access to
cluster data, but a planned revamp of the enrolment logic into the
roleserver will fully integrate this with cluster information.

Change-Id: Icc472a0da302109882c5a6d8b4e124a7b9af4813
Reviewed-on: https://review.monogon.dev/c/monogon/+/422
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>