m/n/c/curator: implement Watching NodesInCluster
This pipes etcd ranged watchers for nodes into a Curator RPC. This is to
be used by systems that need to compile information based on all/some
nodes in the cluster, eg. when building a cluster directory or hosts
file with DNS mappings.

The existence of both NodeInCluster and NodesInCluster could be argued
as unnecessary, and it might make sense to merge NodeInCluster
functionality into NodesInCluster with a filter-by-node-id field. We
should consider doing this once the dust settles.

We also use this opportunity to write tests for Node{,s}InCluster.

Change-Id: I544657b1bfe266a37230760236510024c6007c24
Reviewed-on: https://review.monogon.dev/c/monogon/+/420
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>