m/c/metroctl: use in-tree artifacts if called using bazel run
This allows for not specifying a bundle or installer and instead
taking them from the current build if running under bazel run.
So bazel run //metropolis/cli/metroctl -- install will automatically
build and use the bundle and installer. This just changes the defaults,
if you manually specify installer and/or bundle these are still
respected.

Change-Id: I0676c9b28308544712c06881ad6ace2dba4cab2c
Reviewed-on: https://review.monogon.dev/c/monogon/+/425
Reviewed-by: Leopold Schabel <leo@nexantic.com>