m/node: add a crude MVP bundle
Since the bundle format will likely need some more time cooking let's
just use a quick'n'dirty ZIP file for now. This is explicitly not stable
and will be replaced by the actual bundle format before release. But
it allows us to untangle various parts of the installer machinery and
land them while this is still cooking.

Change-Id: I7ba7875232e4b9a03a4dd564f2ca02d2663f829c
Reviewed-on: https://review.monogon.dev/c/monogon/+/430
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>