m/p/supervisor: deflake tests
This removes some poorly designed test code which attempted to
synchronize with a goroutine in a way that's unsound in Go's concurrency
model. Instead of doing a non-blocking read and failing if there is no
sending goroutine, we just block. Test timeouts will, in this case,
cause the test to error out.

Change-Id: I5338693c578c8eb8b494a1a651a04de6a54df15c
Reviewed-on: https://review.monogon.dev/c/monogon/+/445
Reviewed-by: Leopold Schabel <leo@nexantic.com>