m/n/installer: reboot on fatal errors
Currently irrecoverable errors are handled by log.Fatal, which calls
os.Exit after printing the error message. As this kills the init
process, it leads the kernel to panic and print a long call trace,
obscuring the true cause of trouble. The kernel then reboots
immediately, as it is configured to do so, making it impossible to
read the logs off the screen when using a graphical console.

This patch makes the installer hang after displaying status
information.

Change-Id: Ib6a7582b621fd5072e636242f8279be2a831c3f3
Reviewed-on: https://review.monogon.dev/c/monogon/+/488
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>