metropolis: align EFI partition layout
This moves non-EFI-related data inside the ESP from under the EFI
subdirectory to the root. The data moved is never accessed from EFI and
thus shouldn't be under the EFI folder. This also aligns localstorage
with the new layout as previously it and osimage didn't agree on a
specific layout, indirectly breaking the installer.

Change-Id: I36bdc9782e181dafab40aaa85cc0b4eaf2448f6e
Reviewed-on: https://review.monogon.dev/c/monogon/+/483
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>