m/n/core/network/dhcp4c: fix logic bug in DNSServers
This previously accepted all non-undefined (0.0.0.0) IPs as servers on
account of the `|| ip4Num != 0` in the condition. That's unnecessary
anyways, so let's drop it.
While we're in there I also changed the deduplicating map from a
map[uint32]struct{} to a map[uint32]bool, making the code a bit shorter.

Change-Id: Ic15cb96217a300913ebc58580d4314a6449da923
Reviewed-on: https://review.monogon.dev/c/monogon/+/516
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>