metropolis/node/build: prepare fsspec for use in initramfs
This moves fsspec up into the node build directory as it's going to be
used in multiple tools (like mkinitcpio) in the future. It also adds
special files support to it as that's going to be required at least for
initramfs support but it may prove useful for erofs as well.

Change-Id: I8d559bb761b4da350c0070b23b5ab393ba6e9872
Reviewed-on: https://review.monogon.dev/c/monogon/+/521
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>