m/p/supervisor: implement sub-loggers
This permits logging from a runnable into a logtree sub-DN that is not
backed by an actualy child runnable. For example, 'root.foo' can request
a SubLogger with name 'bar' to emit logs into 'root.foo.bar'.

This is in preparation for logging RPC calls within supervised
runnables, but can also come in handy in other situations where we'd
like to log to separete 'topics' within a single runnable.

This breaks 1:1 correspondence between logtree DNs and supervisor DNs.
An alternative would be to introduce extra 'tags'/'topics' eg
root.foo:bar, but that would require encoding extra logic to the
logtree. However, that would perhaps allow us to introduce higher
cardinality child loggers, with a logger per RPC. We'll have to consider
this at some later point.

Let's see where this takes us, there's a chance we'll roll this
back if it's too confusing from an UX point of view.

Change-Id: Ibdee5c2b400bb8fce76b0a4f781914748793db0e
Reviewed-on: https://review.monogon.dev/c/monogon/+/536
Reviewed-by: Leopold Schabel <leo@nexantic.com>