m/n/core: only run debug service in debug build
This excludes the debug service from non-debug builds as it exposes a
bunch of unauthenticated interfaces for debugging to the world.
The Kubernetes tests were the last user of this service but getting
Kubernetes credentials is now handled by an authenticated production
service (the authproxy).
Some parts of the debug service functionality, namely GetLogs will also
be needed outside of debug builds, but nothing depends on its
availability so we can do this right away.

Change-Id: I5ba3d2853c69ae295d6224b359b36c160b58c430
Reviewed-on: https://review.monogon.dev/c/monogon/+/552
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>