m/p/supervisor: wait for runnables to exit in TestHarness
This ensures that tests which aren't marked as parallel won't interfere
with eachother due to still running runnables (for example, gracefully
terminating gRPC services listening on some stable port number).

To implement this, we add the Liquidator, a goroutine responsible for
maintaining a minimum viable supervisor processor which records all
runnables' exits. These can then be inspected by the TestHarness to
ensure that all runnables are truly dead.

Change-Id: I436f9608d1e0e04796f7198b641e7d625df885f8
Reviewed-on: https://review.monogon.dev/c/monogon/+/625
Reviewed-by: Leopold Schabel <leo@nexantic.com>