m/n/core/rpc: add ClusterResolver
This is a first-pass implementation of a baseline, functioning, but not
fully featured gRPC resolver builder that connects to a given Metropolis
cluster based on just a single functioning node.

This is planned to be extended to be aware of node health, and possibly
curator leadership. It will then replace the main roleserver client and
allow metroctl to connect to a cluster given just a single node.

Change-Id: I8321a6ce19bdaead35b5f266dd9774ce1b78f075
Reviewed-on: https://review.monogon.dev/c/monogon/+/637
Reviewed-by: Lorenz Brun <lorenz@monogon.tech>