m/n/k/authproxy: make use of SPDY through proxy work
Kubernetes still uses SPDY for interactive/streaming-type calls (like
exec or port-forward). Our proxy uses a HTTP/2 backend connection to
Kubernetes's API server. A HTTP/2 stream cannot be upgraded to SPDY
meaning these API requests all fail. This implements a slightly ugly
workaround by using two HTTP transports, a regular transport which
supports HTTP/2 and a fallback transport which does not. The proxy
selects the fallback transport if it detects that the request is trying
to upgrade to SPDY.

Change-Id: Idd44f58d07ec5570ddf8941ae7595225f47f254d
Reviewed-on: https://review.monogon.dev/c/monogon/+/645
Reviewed-by: Sergiusz Bazanski <serge@monogon.tech>